requirements.txt
boto==2.38.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.10.6
django-storages==1.1.8
django-tenant-schemas==1.6.11
django-toolbelt==0.0.1
gunicorn==19.3.0
newrelic==2.64.0.48
psycopg2==2.6.2
static==0.4
static3==0.6.1
